Export build system as node module which modifies an existing instance of gulp
