+++ app_gulpfile.js
@@ -1,0 +1,9 @@
+/**
+ * IMPORTANT: This file is installed into the app directory as part of the
+ * npm install fxos-build process. Please do not try to modify this file.
+ *
+ * For any build system feature requests or bugs, please open
+ * an issue in the fxos-build project: https://github.com/fxos/build/issues
+ */
+var gulp = require('gulp');
+require('fxos-build')(gulp);
